requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
mysql==0.0.3
mysqlclient==2.1.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
